- swagger-codegen 버전 업 (2.2.3 -> 2.3.0) 그런데 소스는 좀 더 지저분해진 느낌.. - 관련하여 코드 좀 더 수정 - skin 설정같은 API에서 오류 발생. 조건은 POST에 parameter가 1개이고, 이 파라미터가 json 변환이 될 수 없는 타입이면, 오류가 발생한다. - codegen 을 올렸더니 DefaultAPI 외에 _API가 하나 더 추가되더라... 저건 나조차도 용납할 수 없는 이름이어서, 일단 DefaultAPI2로 하고 DefaultAPI의 extension으로 확장하였다.

+++ BoostMINI/BoostMINI/Classes/ViewControllers/SettingScreens/SettingViewController.swift
@@ -190,7 +190,7 @@ extension SettingViewController : UICollectionViewDelegate, UICollectionViewData
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         
         if let skins = self.skinDatas {
-            DefaultAPI.postSkinsUsingPOST(select: Int32(indexPath.row), completion: { (response, error) in
+            DefaultAPI.postSkinsUsingPOST(select: indexPath.row, completion: { (response, error) in
                 guard let data = response else {
                     return
                 }
